pyproject.toml
# See the NOTICE file distributed with this work for additional information
# regarding copyright ownership.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
# http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# pyproject.toml -- VEuPathDB Redmine client project configuration

## Generic configuration
[project]
name = "veupath-redmine-client"
dynamic = [
    "version",
]
requires-python = ">= 3.8"
description = "A VEuPathDB specific Redmine client"
readme = "README.md"
authors = [
    {name = "VEuPathDB", email = "help@veupathdb.org"},
]
license = {text = "Apache License 2.0"}
keywords = [
    "redmine",
    "veupathdb",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "biopython",
    "python-redmine",
]

[project.optional-dependencies]
dev = [
    "black",
]

[project.urls]
homepage = "https://www.veupathdb.org"
repository = "https://github.com/VEuPathDB/veupath-redmine-client"

[project.scripts]
# Checks
check_genome_issues = "veupath.redmine.client.checks.check_genome_issues:main"
check_missed_issues = "veupath.redmine.client.checks.check_missed_issues:main"
check_organism_abbrevs = "veupath.redmine.client.checks.check_organism_abbrevs:main"
check_rnaseq_issues = "veupath.redmine.client.checks.check_rnaseq_issues:main"
check_single_issue = "veupath.redmine.client.checks.check_single_issue:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "veupath.redmine.client.__version__"}

[tool.setuptools.packages.find]
where = ["src"]  # list of folders that contain the packages (["."] by default)

# For additional information on `setuptools` configuration see:
#    https://setuptools.pypa.io/en/latest/userguide/quickstart.html
#    https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
#    https://packaging.python.org/en/latest/specifications/declaring-project-metadata/

[build-system]
requires = [
    "setuptools",
    "setuptools-scm",
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 110

[tool.pylint.main]
max-line-length = 110
disable = [
    "invalid-name",
    "logging-fstring-interpolation",
    "similarities",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-locals",
    "too-many-statements",
    "unspecified-encoding",
    "wildcard-import",
]
max-attributes = 12
